Switch resource scoping to use pundit policy scope
We want to implement policies about when a user can edit or delete games. We could do this directly in the resources, but this starts to feel like the resources are gaining a second responsibility. Instead, we'll split scoping and policies into separate objects provided by the Pundit gem.

This commit refactors the current scoping behavior to use Pundit, but does not yet add any policies.